refactor(contracts): extract turn schemas
