+++ packages/contracts/src/index.ts
@@ -1,3 +1,6 @@
 export * from "./app-failure"
+export * from "./chat"
+export * from "./context"
 export * from "./rpc"
 export * from "./streams"
+export * from "./turns"
